+++ sphinx_gallery/tests/test_gen_rst.py
@@ -185,6 +185,19 @@ def test_ipy_notebook():
         f.flush()
         assert_equal(json.load(f), example_nb.work_notebook)
 
+def test_thumbnail_number():
+    # which plot to show as the thumbnail image
+    for test_str in ['# sphinx_gallery_thumbnail_number= 2',
+                     '# sphinx_gallery_thumbnail_number=2',
+                     '#sphinx_gallery_thumbnail_number = 2',
+                     '    # sphinx_gallery_thumbnail_number=2']:
+        with tempfile.NamedTemporaryFile('w') as f:
+            f.write('\n'.join(['"Docstring"',
+                               test_str]))
+            f.flush()
+            _, content = sg.get_docstring_and_rest(f.name)
+            thumbnail_number = sg.extract_thumbnail_number(content)
+        assert_equal(thumbnail_number, 2)
 
 def test_save_figures():
     """Test file naming when saving figures. Requires mayavi."""